feat(Traits) add the With_Caches_Reset trait
